Add AsReadOnly extension for ISet<T>
