+++ src/libraries/System.Collections/src/System.Collections.csproj
@@ -20,7 +20,6 @@
     <Compile Include="System\Collections\Generic\OrderedDictionary.cs" />
     <Compile Include="System\Collections\Generic\PriorityQueue.cs" />
     <Compile Include="System\Collections\Generic\PriorityQueueDebugView.cs" />
-    <Compile Include="System\Collections\Generic\ReadOnlySet.cs" />
     <Compile Include="System\Collections\Generic\SortedDictionary.cs" />
     <Compile Include="System\Collections\Generic\SortedList.cs" />
     <Compile Include="System\Collections\Generic\SortedSet.cs" />
